Reorganizing package structure to generalise metrics reporting
