--- src/main/scala/zio/zmx/metrics/StatsdInstrumentation.scala
@@ -1,9 +1,0 @@
-package zio.zmx.metrics
-
-import zio.ZIO
-import zio.zmx.MetricsDataModel._
-
-private[metrics] class StatsdInstrumentation extends ZMetrics.Service {
-  override def counter(name: String): ZIO[Any, Nothing, Option[Metric.Counter]]        = ???
-  override def increment(m: Metric.Counter): ZIO[Any, Nothing, Option[Metric.Counter]] = ???
-}
