+++ src/main/scala/zio/zmx/statsd/StatsdClient.scala
@@ -1,4 +1,4 @@
-package zio.zmx.metrics
+package zio.zmx.statsd
 
 import zio._
 import zio.zmx.MetricsConfigDataModel._
